requirements.txt
streamlit
sentence-transformers
chromadb
numpy==1.26.4
python-dotenv

# Gemini (PINNED to avoid Windows/Python 3.11 issues)
google-generativeai==0.3.2
google-api-core==2.17.1
googleapis-common-protos==1.63.2
protobuf==4.25.3